Reduce some duplication in stub mapping transformer tests
